Fixing bug with packet crafting
Fixing bug with includes

added build option

Refactored EDO classes

Fixed bug

fix bug

Testing

Testing

Fixed bug with PROTO ID

Finalize EDO Request capability

Patch to add EDO Length TCP Option
