+++ libcrafter/crafter/Crafter.h
@@ -60,7 +60,6 @@ SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 #include "Protocols/TCPOptionPad.h"
 #include "Protocols/TCPOptionMPTCP.h"
 #include "Protocols/TCPOptionExtendedDataOffset.h"
-#include "Protocols/TCPOptionExtendedDataOffsetRequest.h"
 
 /* IPv4 Protocol Implementation */
 #include "Protocols/IP.h"
