--- libcrafter/crafter/ProtoSource/TCPOptionEDOreq.src
@@ -1,4 +1,0 @@
-name TCPOptionExtendedDataOffsetRequest
-protoid 0x0ED0
-ByteField Kind    0 0 0
-ByteField Length  0 1 2
