+++ libcrafter/crafter/Protocols/TCPCraft.cpp
@@ -57,14 +57,16 @@ void TCP::Craft() {
 	if (!IsFieldSet(FieldDataOffset)) {
 		Layer* top_layer = GetTopLayer();
 		size_t option_length = 0;
+		size_t option_edo_ref = 0;
 		TCPOptionExtendedDataOffset* option_edo = NULL;
 
 		if(top_layer) {
 			while( top_layer && ((top_layer->GetID() >> 8) == (TCPOption::PROTO >> 8))) {
-				option_length += top_layer->GetSize();
-				if(top_layer->GetName().compare("TCPOptionExtendedDataOffset")){
+				if(!top_layer->GetName().compare("TCPOptionExtendedDataOffset")){
 					option_edo = (TCPOptionExtendedDataOffset*) top_layer;
+					option_edo_ref = option_length;
 				}
+				option_length += top_layer->GetSize();
 				top_layer = ((TCP *)top_layer)->GetTopLayer();
 			}
 		}
@@ -74,14 +76,14 @@ void TCP::Craft() {
 					     "TCP::Craft()",
 				         "Option size is not padded to a multiple of 4 bytes.");
 		if(!option_edo){
-			SetDataOffset(5 + option_length/4); 	
+			SetDataOffset(5 + option_length/4);
 		}else{
-			/* Check if AO Option (not added yet) */ 
+			/* Check if AO Option (not added yet) */
 			/* Is a padding needed since EDO should be last option related by DataOffset*/
-			SetDataOffset(5 + 1);
+			SetDataOffset(5 + option_edo_ref/4 + 2);
 			option_edo->SetHeader_length(5 + option_length/4);
 		}
-		
+
 		ResetField(FieldDataOffset);
 	}
 
@@ -180,6 +182,7 @@ void TCP::ParseLayerData(ParseInfo* info) {
 		/* Extra information for IP options */
 		TCPOptionLayer::ExtraInfo* extra_info = new TCPOptionLayer::ExtraInfo;
 		extra_info->optlen = TCP_opt_size;
+		extra_info->optlen_origin = TCP_opt_size;
 		extra_info->next_layer = 0;
 
 		/* Information for the decoder */
