--- libcrafter/crafter/Protocols/TCPOptionExtendedDataOffsetRequest.h
@@ -1,84 +1,0 @@
-/*
-Copyright (c) 2012, Esteban Pellegrino
-All rights reserved.
-
-Redistribution and use in source and binary forms, with or without
-modification, are permitted provided that the following conditions are met:
-    * Redistributions of source code must retain the above copyright
-      notice, this list of conditions and the following disclaimer.
-    * Redistributions in binary form must reproduce the above copyright
-      notice, this list of conditions and the following disclaimer in the
-      documentation and/or other materials provided with the distribution.
-    * Neither the name of the <organization> nor the
-      names of its contributors may be used to endorse or promote products
-      derived from this software without specific prior written permission.
-
-THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND
-ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
-WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
-DISCLAIMED. IN NO EVENT SHALL ESTEBAN PELLEGRINO BE LIABLE FOR ANY
-DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES
-(INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
-LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND
-ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
-(INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
-SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
-*/
-#ifndef TCPOPTIONEXTENDEDDATAOFFSETREQUEST_H_
-#define TCPOPTIONEXTENDEDDATAOFFSETREQUEST_H_
-
-#include "../Layer.h"
-
-
-#ifndef TCPOPT_EDOREQ
-#define TCPOPT_EDOREQ   32
-#endif
-
-namespace Crafter {
-
-    class TCPOptionExtendedDataOffsetRequest: public TCPOptionLayer {
-
-        void DefineProtocol();
-
-        Constructor GetConstructor() const {
-            return TCPOptionExtendedDataOffsetRequest::TCPOptionExtendedDataOffsetRequestConstFunc;
-        };
-
-        static Layer* TCPOptionExtendedDataOffsetRequestConstFunc() {
-            return new TCPOptionExtendedDataOffsetRequest;
-        };
-
-        void Craft();
-
-        static const byte FieldKind = 0;
-        static const byte FieldLength = 1;
-
-    public:
-
-        static const word PROTO = 0x0ED0;
-
-        TCPOptionExtendedDataOffsetRequest();
-
-        void SetKind(const byte& value) {
-            SetFieldValue(FieldKind,value);
-        };
-
-        void SetLength(const byte& value) {
-            SetFieldValue(FieldLength,value);
-        };
-
-        byte  GetKind() const {
-            return GetFieldValue<byte>(FieldKind);
-        };
-
-        byte  GetLength() const {
-            return GetFieldValue<byte>(FieldLength);
-        };
-
-        ~TCPOptionExtendedDataOffsetRequest() { /* Destructor */ };
-
-    };
-
-}
-
-#endif /* TCPOPTIONEXTENDEDDATAOFFSETREQUEST_H_ */
